refactor: move getDevices function to createAccountCard
