+++ react-components/src/components/pages/FormPage/functions/createAccountCard.ts
@@ -1,5 +1,4 @@
 import { IAccountCard, IAllInputsData, ICheckboxesData } from 'types';
-import getDevices from './getDevices';
 
 export default function createAccountCard(
   key: number,
@@ -10,6 +9,39 @@ export default function createAccountCard(
   if (inputsData.genderInput) {
     gender = 'Female';
   }
+
+const getDevices = (checkboxesData: ICheckboxesData): string => {
+  let devices = ['none :-('];
+
+  if (
+    checkboxesData.pcCheckbox ||
+    checkboxesData.ps5Checkbox ||
+    checkboxesData.XboxCheckbox ||
+    checkboxesData.switchCheckbox
+  ) {
+    devices = [];
+  }
+
+  if (checkboxesData.pcCheckbox) {
+    devices.push('PC');
+  }
+
+  if (checkboxesData.ps5Checkbox) {
+    devices.push('PS5');
+  }
+
+  if (checkboxesData.XboxCheckbox) {
+    devices.push('Xbox series X');
+  }
+
+  if (checkboxesData.switchCheckbox) {
+    devices.push('Switch');
+  }
+
+  const devicesStr = devices.join(', ');
+
+  return devicesStr;
+}
 
   const devices = getDevices(checkboxesData);
